+++ lib/catcher.dart
@@ -24,3 +24,4 @@ export 'package:catcher/model/report_handler.dart';
 export 'package:catcher/model/report_mode.dart';
 export 'package:catcher/model/toast_handler_gravity.dart';
 export 'package:catcher/model/toast_handler_length.dart';
+export 'package:catcher/utils/catcher_logger.dart';