+++ lib/model/catcher_options.dart
@@ -5,6 +5,7 @@ import 'package:catcher/model/localization_options.dart';
 import 'package:catcher/model/report.dart';
 import 'package:catcher/model/report_handler.dart';
 import 'package:catcher/model/report_mode.dart';
+import 'package:catcher/utils/catcher_logger.dart';
 
 class CatcherOptions {
   /// Handlers that should be used
@@ -50,6 +51,9 @@ class CatcherOptions {
   ///milliseconds.
   final int reportOccurrenceTimeout;
 
+  ///Logger instance.
+  final CatcherLogger? logger;
+
   /// Builds catcher options instance
   CatcherOptions(
     this.reportMode,
@@ -64,6 +68,7 @@ class CatcherOptions {
     this.excludedParameters = const [],
     this.filterFunction,
     this.reportOccurrenceTimeout = 3000,
+    this.logger,
   });
 
   /// Builds default catcher options release instance
@@ -79,7 +84,8 @@ class CatcherOptions {
         screenshotsPath = "",
         excludedParameters = const [],
         filterFunction = null,
-        reportOccurrenceTimeout = 3000;
+        reportOccurrenceTimeout = 3000,
+        logger = CatcherLogger();
 
   /// Builds default catcher options debug instance
   CatcherOptions.getDefaultDebugOptions()
@@ -94,7 +100,8 @@ class CatcherOptions {
         screenshotsPath = "",
         excludedParameters = const [],
         filterFunction = null,
-        reportOccurrenceTimeout = 3000;
+        reportOccurrenceTimeout = 3000,
+        logger = CatcherLogger();
 
   /// Builds default catcher options profile instance
   CatcherOptions.getDefaultProfileOptions()
@@ -109,5 +116,6 @@ class CatcherOptions {
         screenshotsPath = "",
         excludedParameters = const [],
         filterFunction = null,
-        reportOccurrenceTimeout = 3000;
+        reportOccurrenceTimeout = 3000,
+        logger = CatcherLogger();
 }