+++ lib/utils/catcher_utils.dart
@@ -19,4 +19,5 @@ class CatcherUtils {
   static bool isCupertinoAppAncestor(BuildContext context) {
     return context.findAncestorWidgetOfExactType<CupertinoApp>() != null;
   }
+
 }