+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 * Updated dependencies by Nico Mexis (https://github.com/ThexXTURBOXx)
 * Added `reportOccurrenceTimeout` in CatcherOptions to prevent reporting duplicated errors.
 * Added new handler: SnackbarHandler which can be used to show snackbar message.
+* Reworked logger in Catcher. Now Catcher uses only one class to handle log printing. Default logger can be overridden with custom logger.
+* Added `logger` parameter in CatcherOptions to provide custom logger to Catcher.
+
 
 ## 0.6.6
 * Added copyWith method in LocalizationOptions.